* Expanded IFuture interface to allow for cancel and failure commands * Added GetPaused function to determine if a Routine was paused or resumed * Routine.WaitSeconds is now more accurate * Moved Examples into unitypackage
