+++ .gitignore
@@ -30,3 +30,4 @@ sysinfo.txt
 /dist
 /Assets/VSCode
 /Assets/VSCode.meta
+/.vscode
