+++ Assets/BeauRoutine/Examples/1 - Basic/Example1_Script.cs.meta
