+++ Assets/BeauRoutine/Examples/3 - UI Animations.meta
