+++ Assets/BeauRoutine/Examples/3 - UI Animations/Example3_Script.cs.meta
