+++ Assets/BeauRoutine/Examples/3 - UI Animations/Example3_UIAnimations.unity.meta
