+++ Assets/BeauRoutine/Examples/5 - Click the Tweened Button.meta
