+++ Assets/BeauRoutine/Examples/5 - Click the Tweened Button/Example5_ClickTheTweenedButton.unity.meta
